Switched tag to string to the fragment
